requirements.txt
fastapi
uvicorn
requests
aiofiles
python-multipart
psycopg2-binary